fastapi_gateway/requirements.txt
# 의존성 
fastapi
uvicorn
requests
cryptography
pyjwt
